Paxos round number bug fix + improvements.
Previously, Paxos leaders were incrementing their round number by the
number of acceptors instead of the number of leaders. This isn't a bug
when the number of acceptors is larger than the number of leaders, but
is a bug if the number of leaders is large.

This commit fixes that bug and also starts of leaders in round -1 so
that their first round is not really the second round number.