Varying command size and execution time.
The command size implementation is straightforward. The command
execution time is much less straightforward since even a 1ms sleep
greatly affects the performance of the protocol. I probably need to add
much smaller execution times, on the order of microseconds. Also, in a
real implementation, you might defer the execution of the log to a
separate thread so that it doesn't block commands getting chosen.
Though, if you do this, then you can start choosing much faster than you
can execute which is also bad.