[WIP] Progress to fixing EPaxos (without deps).
I'm currently working on fixing the EPaxos implementation with sequence
and dependency computation deferred for later, and with execution
deferred till later. As I was thinking about the protocol, I realized
that EPaxos probably has a bug in the way it implements ballots. I hope
the way we're implementing ballots is correct.