Made proxy leaders non-optional.
Previously, I wanted it to be possible for the leader or the proxy
leader to send phase2a messages to the acceptors. This would allow us to
experiment with and without the proxy leaders. However, this means that
when an acceptor receives a Phase2a, it doesn't know which type of
channel to send back with. It's not a big deal, but it's annoying enough
to just make proxy leaders non-optional. We can always co-locate the two
in a single process to emulate a single leader.