Added spaced flushing to proxy leaders.
I found that with 7 proxy leaders, the proxy leaders were the
bottleneck. Adding spaced flushing helps reduce the number of proxy
leaders we need for them not be the bottleneck.

Now, the leader can send about roughly every 5-6 microseconds or so.
This corresponds to a throughput of 166,666 to 200,000. I'm seeing peaks
of around 175,000 which is right around being bottlenecked by the leader
which is what we want.