First crack at Mencius client.
The only difference here is that the client doesn't have a single leader
to send to. It can send to any of the leaders. One complication is that
re-sending commands upon learning about a new leader becomes a bit more
complicated. For now, we just don't do it. In practice, we don't have
any leader changes.