Slightly better cleanup.
Our benchmark script runs processes on remote hosts. If we kill the
script (i.e. hitting control-c), then we want to make sure the remote
processes get killed too. This is a super janky thing to do. It's very
unpleasant. When I switched over to multiple threads it became even more
unpleasant. It's still not perfect, but hopefully good enough.