[WIP] Started vanilla mencius implementation.
We have a mencius implementation, but it's a compartmentalized version
of Mencius. That version is different enough from Mencius, that I'm just
starting a new implementation.

This commit doesn't build, but the client should be more or less good.
Just need to implement the Server now, which will be a bit of work.