Fixed bad recover timer.
Consider a MultiPaxos replica. It's log may look something like this:

      0   1   2   3   4   5   6   7
    +---+---+---+---+---+---+---+---+
    | x | y | z |   | a |   |   |   |
    +---+---+---+---+---+---+---+---+

Command a is chosen in log entry 4, but it cannot be executed until some
log entry is chosen in log entry 3. However, it's possible that log
entry 3 will never get chosen. For example, the proxy leader that
proposed the command in log entry 3 may have died. If we sit by and do
nothing, we get stuck.

To solve this problem, a replica will detect holes in its log and send a
message to the leader telling it to fill the hole. This is called
recovery. In our implementation, the leader performs a recovery (i.e.
fills a hole) by performing a round change. This is relatively costly,
so we don't want to be filling holes too often.

The code had a small bug in it where the replicas were trying to recover
too aggressively. My intention was to have a replica initiate a recovery
if a hole existed for some period of time. The code unfortunately would
issue a recovery if any hole existed for some period of time.

For example, a replica wants to perform a recovery if a hole exist for 1
second. In the buggy code, if a hole in log entry 3 existed for half a
second and then a hole in log entry 5 existed for half a second, the
replica would initiate the recovery.

This commit fixes this bug. I had already fixed this bug in matchmaker
multipaxos but forgot to port it over.