Increased max message size.
In our Netty transport implementation, a receiver of a message has an
upper bound on the message size it is willing to receive. If it receives
a message that is larger than this upper bound, it ignores it.
Previously, the bound was 1MB, and I was never running into the bound.
However, as I've run benchmarks with higher and higher throughput, I
have sometimes hit this bound. The culprit is big Phase1bs. Sometimes an
acceptor has a lot of (big) log entries for the leader to process, and
the maximum size gets violated.

I solved this by increasing the size to 10 MB. I also added some logging
to better track down who is sending the excessively large messages. I
thought it was the clients for a while before realizing it was the
acceptors.