Cleaned up numBatchers.
Previously, numBatchers was being used as a boolean for whether or not
to batch. Specifically, if numBatchers was 0, then we didn't batch, and
if it was anything other than 0, then we did batch. This numBatchers
parameter is actually unneeded since the client takes in a batchSize
flag. This flag alone determines whether the clients batch or not. If
batchSize == 1, we don't batch. If batchSize > 1, then we do batch.